requirements.txt
flask-wtf
flask_migrate
alembic
flask-migrate
sqlalchemy
pymysql
datetime
flask-login